update and cleanup to use single template
