--- nautilus-gcp-verify/.gitignore
@@ -1,17 +1,0 @@
-*.env
-user-data.sh
-trust-policy.json
-secrets-policy.json
-secrets.json
-config.alloy
-
-/move/enclave/build
-/move/app/build
-
-/src/nautilus-server/target
-
-/out
-/target
-
-/frontend/node_modules
-/frontend/dist
