--- nautilus-gcp-verify/Cargo.toml
@@ -1,13 +1,0 @@
-[workspace]
-members = [
-  "src/aws",
-  "src/init",
-  "src/system"
-]
-
-exclude = [
-  "src/nautilus-server"
-]
-
-# Set default resolver to version 2
-resolver = "2"
