--- nautilus-gcp-verify/README.md
@@ -1,8 +1,0 @@
-# Nautilus: Verifiable offchain computation on Sui
-
-Nautilus is a framework for **secure and verifiable off-chain computation on Sui**. For full product details, see the [Nautilus documentation](https://docs.sui.io/concepts/cryptography/nautilus).
-
-
-```
-curl -X POST http://localhost:3000/verify_google_jwt \\n  -H "Content-Type: application/json" \\n  -d '{"jwt":"eyJhbGciOiJSUzI1NiIsImtpZCI6IjE2ZDdhYmJhMzgxZTNjMWQyNGZmMDA3MjE3NmE0ZDY4ODNkNTNlZjAiLCJ0eXAiOiJKV1QifQ.eyJhdWQiOiJodHRwczovL215LXZlcmlmaWVyLmV4YW1wbGUiLCJleHAiOjE3NTMzOTEyODgsImlhdCI6MTc1MzM4NzY4OCwiaXNzIjoiaHR0cHM6Ly9jb25maWRlbnRpYWxjb21wdXRpbmcuZ29vZ2xlYXBpcy5jb20iLCJuYmYiOjE3NTMzODc2ODgsInN1YiI6Imh0dHBzOi8vd3d3Lmdvb2dsZWFwaXMuY29tL2NvbXB1dGUvdjEvcHJvamVjdHMvazhzLWNsdXN0ZXJzLTM5NzUyMS96b25lcy91cy1jZW50cmFsMS1iL2luc3RhbmNlcy9jb25mc3BhY2UtaGVsbG8tbG92aW5nLWNyYW5lIiwiZWF0X25vbmNlIjoiNWYzMGVmODYtNjUyNi00NDY4LWE5NjctMWJhODI4YTM0ZDQ0IiwiZWF0X3Byb2ZpbGUiOiJodHRwczovL2Nsb3VkLmdvb2dsZS5jb20vY29uZmlkZW50aWFsLWNvbXB1dGluZy9jb25maWRlbnRpYWwtc3BhY2UvZG9jcy9yZWZlcmVuY2UvdG9rZW4tY2xhaW1zIiwic2VjYm9vdCI6dHJ1ZSwib2VtaWQiOjExMTI5LCJod21vZGVsIjoiR0NQX0FNRF9TRVYiLCJzd25hbWUiOiJDT05GSURFTlRJQUxfU1BBQ0UiLCJzd3ZlcnNpb24iOlsiMjUwMzAxIl0sImRiZ3N0YXQiOiJkaXNhYmxlZC1zaW5jZS1ib290Iiwic3VibW9kcyI6eyJjb25maWRlbnRpYWxfc3BhY2UiOnsic3VwcG9ydF9hdHRyaWJ1dGVzIjpbIkxBVEVTVCIsIlNUQUJMRSIsIlVTQUJMRSJdLCJtb25pdG9yaW5nX2VuYWJsZWQiOnsibWVtb3J5IjpmYWxzZX19LCJjb250YWluZXIiOnsiaW1hZ2VfcmVmZXJlbmNlIjoidXMtY2VudHJhbDEtZG9ja2VyLnBrZy5kZXYvazhzLWNsdXN0ZXJzLTM5NzUyMS9jb25mc3BhY2UtdGVzdC9oZWxsb193b3JsZDp2MSIsImltYWdlX2RpZ2VzdCI6InNoYTI1Njo5N2M5ZGZjNmZkNDhiNGIzNDMxZmIwMmU2YzM1ZDNmM2Y1YWFiNWI5ZDkwYmM5NmJjZjQwNDBkYTFmOWMzYjlmIiwicmVzdGFydF9wb2xpY3kiOiJOZXZlciIsImltYWdlX2lkIjoic2hhMjU2OmM3NzdmYTAzZWJmOGYwZDI5OTk4MWM1ZGVkMGU3YTQ1YTdlYmQ1NTAyYmQ4MjI0MDhkZTdiMzU1ZWY3NTY5ZDciLCJlbnYiOnsiSE9TVE5BTUUiOiJjb25mc3BhY2UtaGVsbG8tbG92aW5nLWNyYW5lIiwiUEFUSCI6Ii91c3IvbG9jYWwvc2JpbjovdXNyL2xvY2FsL2JpbjovdXNyL3NiaW46L3Vzci9iaW46L3NiaW46L2JpbiJ9LCJhcmdzIjpbIi9oZWxsb193b3JsZCJdfSwiZ2NlIjp7InpvbmUiOiJ1cy1jZW50cmFsMS1iIiwicHJvamVjdF9pZCI6Ims4cy1jbHVzdGVycy0zOTc1MjEiLCJwcm9qZWN0X251bWJlciI6IjY2MDg5MjIyNzQwNyIsImluc3RhbmNlX25hbWUiOiJjb25mc3BhY2UtaGVsbG8tbG92aW5nLWNyYW5lIiwiaW5zdGFuY2VfaWQiOiI3NzgyNTQ4OTQzMjAzODU5NTU4In19LCJnb29nbGVfc2VydmljZV9hY2NvdW50cyI6WyJjb25mc3BhY2UtcnVubmVyQGs4cy1jbHVzdGVycy0zOTc1MjEuaWFtLmdzZXJ2aWNlYWNjb3VudC5jb20iXX0.zIj35yjvk-sbK8h1xp1lTKqY3wpvGmPbAvsEXGPoR15Rn4Uykj0kITUE_S8UGJ9ijEivCWJ5Da_42GRfP6iBBQyDXpgnHsYB5hVT2D5qOhyID8Usjaz9XThREy7CCtb5Emil89mjhMv8HecX73nlUpvW1yWgFNDvof8jrGNjZDHQl_jfCz4eDaTr20cXdMn7q0vwluSY1rNsvYt5nUZK9tg_UQSGrQQZjq1OUJMUxQh670hmjMqISdx0CvA2wmdeLrCm6_war8dHr57rpl7oQc0s6b3iGMDCmYLPOyv75StqtajsmM9w1WHMIV44-xy0KmwdMr883xq7FetYEPYuoQ"}' | jq
-```
