--- nautilus-gcp-verify/expose_enclave.sh
@@ -1,14 +1,0 @@
-# Copyright (c), Mysten Labs, Inc.
-# SPDX-License-Identifier: Apache-2.0
-#!/bin/bash
-
-# Gets the encalve id and CID
-# expects there to be only one enclave running
-ENCLAVE_ID=$(nitro-cli describe-enclaves | jq -r ".[0].EnclaveID")
-ENCLAVE_CID=$(nitro-cli describe-enclaves | jq -r ".[0].EnclaveCID")
-
-sleep 5
-# Secrets-block
-
-cat secrets.json | socat - VSOCK-CONNECT:$ENCLAVE_CID:7777
-socat TCP4-LISTEN:3000,reuseaddr,fork VSOCK-CONNECT:$ENCLAVE_CID:3000 &
