--- nautilus-gcp-verify/move/app/Move.lock
@@ -1,49 +1,0 @@
-# @generated by Move, please check-in and do not edit manually.
-
-[move]
-version = 3
-manifest_digest = "A868E926BA8101A2D046231171C2BD0DB551A31B660B2594BBCA80E0DCAEEAFF"
-deps_digest = "3C4103934B1E040BB6B23F1D610B4EF9F2F1166A50A104EADCF77467C004C600"
-dependencies = [
-  { id = "Sui", name = "Sui" },
-  { id = "enclave", name = "enclave" },
-]
-
-[[move.package]]
-id = "MoveStdlib"
-source = { git = "https://github.com/MystenLabs/sui.git", rev = "framework/testnet", subdir = "crates/sui-framework/packages/move-stdlib" }
-
-[[move.package]]
-id = "Sui"
-source = { git = "https://github.com/MystenLabs/sui.git", rev = "framework/testnet", subdir = "crates/sui-framework/packages/sui-framework" }
-
-dependencies = [
-  { id = "MoveStdlib", name = "MoveStdlib" },
-]
-
-[[move.package]]
-id = "enclave"
-source = { local = "../enclave" }
-
-dependencies = [
-  { id = "Sui", name = "Sui" },
-]
-
-[move.toolchain-version]
-compiler-version = "1.46.1"
-edition = "2024.beta"
-flavor = "sui"
-
-[env]
-
-[env.devnet]
-chain-id = "c2bef0b2"
-original-published-id = "0x73736a90ab670cc6f6e6b37c173059dbe5973cd425865195c1be41d6fcaf5c43"
-latest-published-id = "0x73736a90ab670cc6f6e6b37c173059dbe5973cd425865195c1be41d6fcaf5c43"
-published-version = "1"
-
-[env.testnet]
-chain-id = "4c78adac"
-original-published-id = "0x7e712fd9e5e57d87137440cfea77dc7970575a5c3229d78bb7176ab984d94adf"
-latest-published-id = "0x7e712fd9e5e57d87137440cfea77dc7970575a5c3229d78bb7176ab984d94adf"
-published-version = "1"
