--- nautilus-gcp-verify/move/enclave/Move.lock
@@ -1,40 +1,0 @@
-# @generated by Move, please check-in and do not edit manually.
-
-[move]
-version = 3
-manifest_digest = "C26A901C9CB3E83C8F0117B8D7923EEE6F95561854C8DD43EE9CC59A6B27E475"
-deps_digest = "F8BBB0CCB2491CA29A3DF03D6F92277A4F3574266507ACD77214D37ECA3F3082"
-dependencies = [
-  { id = "Sui", name = "Sui" },
-]
-
-[[move.package]]
-id = "MoveStdlib"
-source = { git = "https://github.com/MystenLabs/sui.git", rev = "framework/testnet", subdir = "crates/sui-framework/packages/move-stdlib" }
-
-[[move.package]]
-id = "Sui"
-source = { git = "https://github.com/MystenLabs/sui.git", rev = "framework/testnet", subdir = "crates/sui-framework/packages/sui-framework" }
-
-dependencies = [
-  { id = "MoveStdlib", name = "MoveStdlib" },
-]
-
-[move.toolchain-version]
-compiler-version = "1.46.1"
-edition = "2024.beta"
-flavor = "sui"
-
-[env]
-
-[env.testnet]
-chain-id = "4c78adac"
-original-published-id = "0x14e8b4d8b28ee9aa5ea604f3f33969b3d0f03247b51837f27e17bcf875d3582c"
-latest-published-id = "0x14e8b4d8b28ee9aa5ea604f3f33969b3d0f03247b51837f27e17bcf875d3582c"
-published-version = "1"
-
-[env.devnet]
-chain-id = "c2bef0b2"
-original-published-id = "0x5baaf9849ff54513f6021d4a758fbb4cdb794eb344045d023b79e9c7d5a0d4b5"
-latest-published-id = "0x5baaf9849ff54513f6021d4a758fbb4cdb794eb344045d023b79e9c7d5a0d4b5"
-published-version = "1"
