--- nautilus-gcp-verify/reset_enclave.sh
@@ -1,3 +1,0 @@
-sudo nitro-cli terminate-enclave --all
-sudo killall socat
-#
