--- nautilus-gcp-verify/rust-toolchain
@@ -1,1 +1,0 @@
-1.81.0
