--- nautilus-gcp-verify/src/aws/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "aws"
-version = "0.1.0"
-edition = "2021"
-license = "ISC"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-libc = "0.2.134"
-nsm_lib = { git = "https://github.com/aws/aws-nitro-enclaves-nsm-api.git/", rev = "8ec7eac72bbb2097f1058ee32c13e1ff232f13e8", package="nsm-lib", optional = false }
-nsm_api = { git = "https://github.com/aws/aws-nitro-enclaves-nsm-api.git/", rev = "8ec7eac72bbb2097f1058ee32c13e1ff232f13e8", package="aws-nitro-enclaves-nsm-api", optional = false }
-system = { path = "../system"}
