--- nautilus-gcp-verify/src/init/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "init"
-version = "0.1.0"
-edition = "2021"
-license = "ISC"
-
-[dependencies]
-libc = "0.2.134"
-aws = { path = "../aws"}
-system = { path = "../system"}
-
-
-[[bin]]
-name = "init"
-path = "init.rs"
