--- nautilus-gcp-verify/src/nautilus-server/Cargo.toml
@@ -1,29 +1,0 @@
-[package]
-name = "nautilus-server"
-version = "0.1.0"
-edition = "2021"
-authors = ["Mysten Labs <build@mystenlabs.com>"]
-license = "Apache-2.0"
-repository = "https://github.com/MystenLabs/nautilus"
-
-[workspace]
-
-[dependencies]
-serde_json = "1.0.140"
-serde_bytes = "0.11"
-serde = "1.0"
-serde_repr = "0.1"
-
-tokio = { version = "1.43.0", features = ["full"] }
-tracing = "0.1"
-axum = { version = "0.7", features = ["macros"] }
-rand = "0.8.5"
-reqwest = { version = "0.11", features = ["json"] }
-anyhow = "1.0"
-serde_yaml = "0.9.34"
-tower-http = { version = "0.6.0", features = ["cors"] }
-fastcrypto = { git = "https://github.com/MystenLabs/fastcrypto", rev = "69d496c71fb37e3d22fe85e5bbfd4256d61422b9", features = ["aes"] }
-nsm_api = { git = "https://github.com/aws/aws-nitro-enclaves-nsm-api.git/", rev = "8ec7eac72bbb2097f1058ee32c13e1ff232f13e8", package="aws-nitro-enclaves-nsm-api", optional = false }
-bcs = "0.1.6"
-jsonwebtoken = "9"
-base64 = "0.22"
