--- nautilus-gcp-verify/src/nautilus-server/allowed_endpoints.yaml
@@ -1,3 +1,0 @@
-# External endpoints that the enclave is allowed to access.
-endpoints:
-  - www.googleapis.com # For fetching JWKS from Google
