--- nautilus-gcp-verify/src/system/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "system"
-version = "0.1.0"
-edition = "2021"
-license = "ISC"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-libc = "0.2.134"
