+++ src/nautilus-server/src/lib.rs
@@ -8,6 +8,13 @@ use axum::Json;
 use fastcrypto::ed25519::Ed25519KeyPair;
 use serde_json::json;
 
+#[cfg(feature = "gcp-verify-example")]
+use std::collections::HashMap;
+#[cfg(feature = "gcp-verify-example")]
+use std::sync::Arc;
+#[cfg(feature = "gcp-verify-example")]
+use tokio::sync::RwLock;
+
 mod apps {
     #[cfg(feature = "twitter-example")]
     #[path = "twitter-example/mod.rs"]
@@ -16,6 +23,10 @@ mod apps {
     #[cfg(feature = "weather-example")]
     #[path = "weather-example/mod.rs"]
     pub mod weather_example;
+
+    #[cfg(feature = "gcp-verify-example")]
+    #[path = "gcp-verify-example/mod.rs"]
+    pub mod gcp_verify_example;
 }
 
 pub mod app {
@@ -24,9 +35,24 @@ pub mod app {
 
     #[cfg(feature = "weather-example")]
     pub use crate::apps::weather_example::*;
+
+    #[cfg(feature = "gcp-verify-example")]
+    pub use crate::apps::gcp_verify_example::*;
 }
 
 pub mod common;
+
+/// JWKS key structure for JWT verification
+#[cfg(feature = "gcp-verify-example")]
+#[derive(Debug, Clone)]
+pub struct JwksKey {
+    pub kid: String,
+    pub alg: String,
+    pub kty: String,
+    pub use_: Option<String>,
+    pub n: String,
+    pub e: String,
+}
 
 /// App state, at minimum needs to maintain the ephemeral keypair.  
 pub struct AppState {
@@ -34,6 +60,9 @@ pub struct AppState {
     pub eph_kp: Ed25519KeyPair,
     /// API key when querying api.weatherapi.com
     pub api_key: String,
+    /// JWKS cache for JWT verification (only for gcp-verify-example)
+    #[cfg(feature = "gcp-verify-example")]
+    pub jwks_cache: Option<Arc<RwLock<(HashMap<String, JwksKey>, u64)>>>,
 }
 
 /// Implement IntoResponse for EnclaveError.
